requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
pyyaml
allure-pytest
selenium
pywinauto